app entry point now main.js. renamed app.js
